Align mlxconfig and tools layouts,  VR branch from mft to mstflint.
Description:
include internal change-ids:
I83ad6d87b1e13cae7e5d4ff01387432b9f32a075
Ic814412f84882dcff370e041ef0d199263b8483c
I0142257c66a6aae5a1e9e56c35d9b1a24efd4965

and other small changes

+++ tools_layouts/adb_to_c_utils.h
@@ -73,7 +73,7 @@ extern "C"
 
 /* MASK generate a bit mask S bits width */
 //#define ADB2C_MASK32(S)     ( ((u_int32_t) ~0L) >> (32-(S)) )
-#define ADB2C_MASK8(S) (((u_int8_t)~0) >> (8 - (S)))
+#define ADB2C_MASK8(S) ((u_int32_t)(((u_int8_t)~0) >> (8 - (S))))
 
 /* BITS generate a bit mask with bits O+S..O set (assumes 32 / 8 bit integer) */
 //#define ADB2C_BITS32(O,S)   ( ADB2C_MASK32(S) << (O) )
